AdminBeneficiariesListTable renamed from table.
